pyproject.toml
[tool.poetry]
name = "hh-api-project"
version = "0.1.0"
description = ""
authors = ["dudeslav1ch <dudeslav1ch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-cov = "^4.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
